# specs/openapi.yaml
openapi: '3.0.0'
info:
  title: 'lenreg_ticket_bot — Архитектура системы'
  description: >
    Единый источник истины (SSOT) для архитектуры данных, бизнес-логики
    и внешних интеграций Telegram-бота мониторинга талонов zdrav.lenreg.ru.

    **Назначение:** отслеживание появления свободных талонов к врачам
    в поликлиниках Ленинградской области через портал zdrav.lenreg.ru
    с уведомлениями в Telegram.

    **Стек:** Python 3.12+, aiogram 3.x, aiosqlite (SQLite WAL),
    httpx (async HTTP), pydantic-settings, Loguru.

    **Архитектурный стиль:** многослойная (handlers → services → database),
    событийно-ориентированная (aiogram), с фоновыми asyncio-циклами.
  version: '1.0.1'
  contact:
    name: 'lenreg_ticket_bot'
    email: 'maintainer@example.com'
    url: 'https://github.com/acidmsg/lenreg_ticket_bot'
servers:
  - url: 'https://zdrav.lenreg.ru/api'
    description: 'Публичный API портала zdrav.lenreg.ru'
...
paths:
...
  /api/dashboard/users/{user_id}:
    get:
      tags:
        - 'Веб-дашборд'
      summary: 'Детали пользователя'
      description: >
        Возвращает полную информацию о пользователе: пациенты, мониторинг,
        последние сообщения. Реализация [`api_user_detail`](src/web/routers/api.py:143).
      operationId: dashboardUserDetail
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Детали пользователя'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '404':
          description: 'Пользователь не найден'
...
  /api/dashboard/logs:
    get:
      tags:
        - 'Веб-дашборд'
      summary: 'Лог мониторинга'
      description: >
        Возвращает записи лога мониторинга с пагинацией и фильтрацией
        по uid и status. Реализация [`api_logs`](src/web/routers/api.py:164).
      operationId: dashboardLogs
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 500
        - name: uid
          in: query
          schema:
            type: string
            nullable: true
        - name: status
          in: query
          schema:
            type: string
            nullable: true
      responses:
        '200':
          description: 'Лог мониторинга'
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/LogEntry'
                  total:
                    type: integer
                  offset:
                    type: integer
                  limit:
                    type: integer
...
components:
  securitySchemes:
    CsrfTokenAuth:
      type: apiKey
      in: header
      name: X-CSRFToken
      description: 'CSRF-токен (статический: NOTPROVIDED)'
...
  schemas:
...
    UserInfo:
      type: object
      description: 'Информация о пользователе Telegram'
      properties:
        uid:
          type: string
          description: 'ID пользователя Telegram'
        patient_count:
          type: integer
          description: 'Количество пациентов пользователя'
        monitoring_count:
          type: integer
          description: 'Количество отслеживаемых врачей'
        last_activity_ts:
          type: number
          description: 'Unix-timestamp последней активности'
        patients:
          type: object
          description: 'Словарь пациентов (только в детальном ответе)'
          additionalProperties:
            $ref: '#/components/schemas/DB_Patient'
        monitoring:
          type: object
          description: 'Словарь мониторинга (только в детальном ответе)'
        last_messages:
          type: object
          description: 'Словарь последних сообщений (только в детальном ответе)'
...
    LogEntry:
      type: object
      description: 'Запись лога мониторинга (для веб-дашборда)'
      properties:
        id:
          type: integer
          description: 'ID записи'
        uid:
          type: string
          description: 'ID пользователя'
        patient_name:
          type: string
          description: 'ФИО пациента'
        doctor_name:
          type: string
          description: 'ФИО врача'
        specialty:
          type: string
          description: 'Специальность'
        clinic_name:
          type: string
          description: 'Название клиники'
        slot_date:
          type: string
          description: 'Дата слота'
        status:
          type: string
          description: 'Статус изменения'
        ts:
          type: number
          description: 'Unix-timestamp события'
...
    DB_Patient:
      type: object
      description: 'Запись таблицы `user_patients`'
      properties:
        fio:
          type: string
          description: 'ФИО пациента'
        bday:
          type: string
          description: 'Дата рождения (YYYY-MM-DD)'
        alias:
          type: string
          nullable: true
          description: 'Псевдоним (опционально, до 25 символов)'
        clinic_id:
          type: string
          nullable: true
          description: 'ID клиники по умолчанию (используется как fallback в мониторинге)'
        confirmed_clinics:
          type: array
          items:
            type: integer
          description: 'Список ID клиник, подтверждённых пользователем'
...